# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiErrorMessage {
  error: string;
}

export type AssetMode = (typeof AssetMode)[keyof typeof AssetMode];

export const AssetMode = {
  vehicles: "vehicles",
  baggage: "baggage",
  cloakrooms: "cloakrooms",
  bags: "bags",
} as const;

export type AssetStatus = (typeof AssetStatus)[keyof typeof AssetStatus];

export const AssetStatus = {
  active: "active",
  released: "released",
} as const;

export interface Patron {
  name: string;
  phone: string;
}

export type AssetFieldValue = string | number | boolean;

export type CustodyAssetFields = { [key: string]: AssetFieldValue };

export interface CustodyAsset {
  id: string;
  ticketId: string;
  venueCode: string;
  mode: AssetMode;
  patron: Patron;
  fields: CustodyAssetFields;
  photos: string[];
  /** Epoch milliseconds */
  intakeAt: number;
  handler: string;
  status: AssetStatus;
  /**
   * Epoch milliseconds, present when released
   * @nullable
   */
  releasedAt?: number | null;
  /** HMAC signature of the ticket id, used by handler-issued QR codes */
  signature: string;
}

export type CreateAssetRequestFields = { [key: string]: AssetFieldValue };

export interface CreateAssetRequest {
  mode: AssetMode;
  patron: Patron;
  fields: CreateAssetRequestFields;
  photos: string[];
  handlerEmail: string;
  handlerName: string;
  venueName?: string;
}

/**
 * Where the ticket id originated. "scan" requires a valid signature; "manual" is the fallback when a handler types the id.
 */
export type ReleaseAssetRequestSource =
  (typeof ReleaseAssetRequestSource)[keyof typeof ReleaseAssetRequestSource];

export const ReleaseAssetRequestSource = {
  scan: "scan",
  manual: "manual",
} as const;

export interface ReleaseAssetRequest {
  handlerEmail?: string;
  handlerName?: string;
  /** HMAC signature from a scanned QR; verified server-side. Omit for manual entry fallback. */
  signature?: string;
  /** Where the ticket id originated. "scan" requires a valid signature; "manual" is the fallback when a handler types the id. */
  source?: ReleaseAssetRequestSource;
}