# lib/api-zod/src/generated/types/custodyAsset.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AssetMode } from "./assetMode";
import type { AssetStatus } from "./assetStatus";
import type { CustodyAssetFields } from "./custodyAssetFields";
import type { Patron } from "./patron";

export interface CustodyAsset {
  id: string;
  ticketId: string;
  venueCode: string;
  mode: AssetMode;
  patron: Patron;
  fields: CustodyAssetFields;
  photos: string[];
  /** Epoch milliseconds */
  intakeAt: number;
  handler: string;
  status: AssetStatus;
  /**
   * Epoch milliseconds, present when released
   * @nullable
   */
  releasedAt?: number | null;
  /** HMAC signature of the ticket id, used by handler-issued QR codes */
  signature: string;
}